+++ product-service/src/main/java/com/example/product_service/entity/Product.java
@@ -18,5 +18,5 @@ public class Product {
     private String name;
     private int price;
     private int quantity;
-    private int orderId;
+    private long orderId;
 }
